Move dispatch into host/dispatch.ts; reduce host.ts to a pure barrel
Move handleMessage/dispatchMessage/handleOp/handleTool/cleanupPort into
host/dispatch.ts. host.ts is now a pure re-export barrel over the host/ family
modules (mirroring client.ts over client/), exposing the exact same public
surface: ensureAuth/portSession, HostCtx/PortLike, the instance-id + branch
symbols, and handleMessage/cleanupPort. Export surface verified byte-identical.
No behavior change.
